refactor: rename epoll to client
